macOS: Workaround overlinking / ld64 not being good at stripping, add libiconv + bzip2 deps
We probably need to use -dead_strip_dylibs and -no_implicit_dylibs in the clang activate script,
but I tested with both those and still got overlinking, so it's probably a bug in ld64. Either
way I do not have time to fix this at present so instead I hack a Makefile.